analytics_vm/requirements.txt
pandas
psycopg2-binary
google-cloud-storage
sqlalchemy
